refactor: organize unit tests in directory
